Improve test framework on updating and debugging
  * test.sh: Stop testing if build fails. For all the test cases, gene-
    rate the correct output on the fly.

+++ test.sh
@@ -3,16 +3,19 @@
 ##** test.sh - A simple test framework ********************************
 
 make clean
-make
-for x in samples/*.decaf
-do
-    ./dcc < $x &> $x.test
-    if diff -urw $x.test ${x/.decaf/.out} > /dev/null
-    then
-        echo -e "\e[32m${x}\e[0m"
-    else
-        echo -e "\e[31m${x}\e[0m"
-        diff -urw $x.test ${x/.decaf/.out}
-    fi
-done
-rm samples/*.decaf.test
+if make
+then
+    for x in samples/*.decaf
+    do
+        ./solution/dcc < $x &> ${x/.decaf/.out}
+        ./dcc < $x &> $x.test
+        if diff -urw $x.test ${x/.decaf/.out} > /dev/null
+        then
+            echo -e "\e[32m${x}\e[0m"
+        else
+            echo -e "\e[31m${x}\e[0m"
+            diff -urw $x.test ${x/.decaf/.out}
+        fi
+    done
+    rm samples/*.decaf.test
+fi
